+++ sample/shared/src/androidMain/kotlin/com/hoc081098/kmpviewmodelsample/common/Immutable.kt
@@ -1,4 +1,0 @@
-package com.hoc081098.kmpviewmodelsample.common
-
-@Suppress("ACTUAL_WITHOUT_EXPECT") // TODO: https://youtrack.jetbrains.com/issue/KT-37316
-internal actual typealias Immutable = androidx.compose.runtime.Immutable